requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.10.0
django-cors-headers==4.7.0
django-filter==25.1
django-modeltranslation==0.19.17
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
idna==3.10
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1

